services/cntk-image-recon/requirements.txt
grpcio>=1.14.2
grpcio-tools>=1.14.1
requests==2.20.0
Pillow==6.2.2
cntk-gpu==2.7